Refactor scripts into ES modules
