+++ contact.html
@@ -58,6 +58,6 @@
     </footer>
   </main>
   <script src="https://cdn.jsdelivr.net/npm/locomotive-scroll@4.1.4/dist/locomotive-scroll.min.js"></script>
-  <script src="script.js"></script>
+  <script type="module" src="main.js"></script>
 </body>
 </html>
