+++ index.html
@@ -77,6 +77,6 @@
 
   </main>
   <script src="https://cdn.jsdelivr.net/npm/locomotive-scroll@4.1.4/dist/locomotive-scroll.min.js"></script>
-  <script src="script.js"></script>
+  <script type="module" src="main.js"></script>
 </body>
 </html>
